+++ libs/targets/gesha-rust-types/src/lib.rs
@@ -10,6 +10,9 @@ pub use definitions::Definitions;
 mod derive_attribute;
 pub use derive_attribute::DeriveAttribute;
 
+mod doc_comments;
+pub use doc_comments::DocComments;
+
 mod enum_variant;
 pub use enum_variant::{EnumCase, EnumVariant, EnumVariantAttribute};
 
@@ -28,6 +31,9 @@ pub use media_type_def::MediaTypeDef;
 mod modules;
 pub use modules::Modules;
 
+mod module_name;
+pub use module_name::ModuleName;
+
 mod request_body_def;
 pub use request_body_def::{MediaTypeVariant, MediaTypeVariants, RequestBodyDef};
 
@@ -42,38 +48,3 @@ pub use struct_field_name::StructFieldName;
 
 mod type_header;
 pub use type_header::TypeHeader;
-
-use std::fmt::{Debug, Display, Formatter};
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq)]
-pub struct ModuleName(String);
-
-impl ModuleName {
-    pub fn new<A: Into<String>>(a: A) -> Self {
-        Self(a.into())
-    }
-}
-
-impl Display for ModuleName {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        Display::fmt(&self.0, f)
-    }
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq)]
-pub struct DocComments(Option<String>);
-
-impl DocComments {
-    pub fn wrap(this: Option<String>) -> Self {
-        Self(this.map(|text| format!("/**\n{text}\n*/\n")))
-    }
-}
-
-impl Display for DocComments {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match &self.0 {
-            Some(text) => Display::fmt(text, f),
-            None => Ok(()),
-        }
-    }
-}
