+++ libs/targets/gesha-rust-types/src/module_name.rs
@@ -1,0 +1,16 @@
+use std::fmt::{Display, Formatter};
+
+#[derive(Clone, Debug, Hash, Eq, PartialEq)]
+pub struct ModuleName(String);
+
+impl ModuleName {
+    pub fn new<A: Into<String>>(a: A) -> Self {
+        Self(a.into())
+    }
+}
+
+impl Display for ModuleName {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        Display::fmt(&self.0, f)
+    }
+}
